feat: make flagged taskPagination code production-ready
