+++ src/shared/containers/SetOrg.tsx
@@ -9,10 +9,8 @@ import PageSpinner from 'src/perf/components/PageSpinner'
 import {
   MePage,
   TasksPage,
-  TasksPagePaginated,
   TaskPage,
   TaskRunsPage,
-  TaskRunsPagePaginated,
   TaskEditPage,
   DashboardsIndex,
   DashboardsIndexPaginated,
@@ -157,25 +155,10 @@ const SetOrg: FC = () => {
           <Route path={`${orgPath}/checks/:checkID`} component={CheckHistory} />
 
           {/* Tasks */}
-          {isFlagEnabled('paginatedTasks') ? (
-            <Route
-              path={`${orgPath}/tasks/:id/runs`}
-              component={TaskRunsPagePaginated}
-            />
-          ) : (
-            <Route
-              path={`${orgPath}/tasks/:id/runs`}
-              component={TaskRunsPage}
-            />
-          )}
+          <Route path={`${orgPath}/tasks/:id/runs`} component={TaskRunsPage} />
           <Route path={`${orgPath}/tasks/:id/edit`} component={TaskEditPage} />
           <Route path={`${orgPath}/tasks/new`} component={TaskPage} />
-          {isFlagEnabled('paginatedTasks') ? (
-            <Route path={`${orgPath}/tasks`} component={TasksPagePaginated} />
-          ) : (
-            <Route path={`${orgPath}/tasks`} component={TasksPage} />
-          )}
-
+          <Route path={`${orgPath}/tasks`} component={TasksPage} />
           {/* Data Explorer */}
           <Route
             path={`${orgPath}/data-explorer`}
